pyproject.toml
[tool.poetry]
name = "url-shortener"
version = "0.1.0"
description = "A URL shortener service built with Django and DRF"

authors = ["Filip Mlynarski <filmlynarski@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.7"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.9"
gunicorn = "^21.2.0"
whitenoise = "^6.6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-django = "^4.5.2"
ruff = "^0.1.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "W", "N", "B", "C4", "UP", "ANN", "RUF"]
ignore = ["ANN", "RUF012"]
exclude = ["*/migrations/*.py"]

[tool.ruff.isort]
known-first-party = ["url_shortener", "shortener"]

[tool.ruff.flake8-quotes]
docstring-quotes = "double"
inline-quotes = "single"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "url_shortener.settings"
python_files = ["test_*.py", "*_test.py"]
